requirements.txt
aenum==3.1.16
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.10.0
attrs==25.4.0
Authlib==1.6.8
backoff==2.2.1
beautifulsoup4==4.14.3
browser-use==0.11.12
browser-use-sdk==2.0.15
bubus==1.5.6
cdp-use==1.4.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.2
cryptography==46.0.5
Cython==3.2.4
Deprecated==1.3.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.8
frozenlist==1.8.0
future==1.0.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-genai==1.64.0
googleapis-common-protos==1.72.0
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
inquirerpy==0.3.4
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-core==1.2.15
langchain-openai==1.1.10
langsmith==0.7.6
line-bot-sdk==3.22.0
lxml==6.0.2
markdown-it-py==4.0.0
markdownify==1.2.2
MarkupSafe==3.0.2
mcp==1.26.0
mdurl==0.1.2
multidict==6.7.1
oauthlib==3.3.1
ollama==0.6.1
openai==2.23.0
orjson==3.11.3
packaging==26.0
pfzy==0.3.4
pillow==12.1.1
portalocker==2.10.1
posthog==7.9.3
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.11.7
pydantic-settings==2.13.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.11.0
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyotp==2.9.0
pyparsing==3.3.2
pypdf==6.7.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.37.0
regex==2026.2.19
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.0
rpds-py==0.30.0
rsa==4.9.1
screeninfo==0.8.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sse-starlette==3.2.0
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
uuid7==0.1.0
uuid_utils==0.14.1
uv==0.8.11
uvicorn==0.41.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.6.0
websockets==15.0.1
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
